+++ dev/changelog.md
@@ -1,7 +1,7 @@
 # ScottPlot Changelog
 
 ## ScottPlot 4.0.38 ⚠️ _in active development_
-* nothing yet...
+* Improved `Plot.PlotFillAboveBelow()` rendering of data with a non-zero baseline (#477) _Thanks @el-aasi_
 
 ## ScottPlot 4.0.37
 * Fixed a long-running issue related to strong assembly versioning that caused the WPF control to fail to render in the Visual Studio designer in .NET Framework (but not .NET Core) projects (#473, #466, #356) _Thanks @bhairav-thakkar, @riquich, @Helitune-RobMcKay, and @iu2kxv_